Update v4 modules for the latest changes in git.
- Remove identmeta (this is now stored by the core).
- Rename forceident to forceusername.
- Fix various other modules.

+++ 4/m_clientcheck.cpp
@@ -105,7 +105,7 @@ public:
 			}
 
 			ServerInstance->Logs.Debug(MODNAME, "Client check: {} -> {} ({})",
-				ci.pattern->GetPattern(), ci.action, ci.message);
+				ci.pattern->GetPattern(), (int)ci.action, ci.message);
 			newclients.push_back(ci);
 		}
